Refactor so that project identifier is better owned
Summary:
The project identifier is constant for any given pyre process - it is
entirely determined by the startup flags and the current directory (i.e.
by CommandArguments), and cannot be changed by mutating the
`.pyre_configuration`.

This matters because we control the ocaml daemon's choice of socket
path using `StartArguments`, which is immutable for a given run, but
in many places we also get it from `frontend_confguration.Base` which
can change during a language server lifetime if the `.pyre_configuration`
changes.

That's unfortunate. The desire to use `Base` to get the project identifier
in some places is driven by the (correct) notion that it isn't clean to be
passing raw arguments dataclasses down to very low layers of the code.
But, it creates the illusion that the socket path might change on us due
to configuration reloading, which is not true (and if it were, the langauge
server would not work!)

We can at least get closer to a clean state by restricting project_identifier
derivations to exactly two places:
- in the construction of a Configuration, where it's pretty easy to see
  that it does not depend on any mutable field, and
- in the construction of the backend start arguments (start.Arguments)
  from a StartArguments dataclass

We need these two to agree, which they do; by removing as many layers
as possible it's much easier to see this than before.

Once that is done, it makes much more sense for daemon_socket to
directly take the project_identifier which is what we do here.

Reviewed By: joshuaxu1995, grievejia

Differential Revision: D39071453

fbshipit-source-id: abf75b857eda3abc161e96470fb3c97bf00c7d81