Push the initial server options read up into run_persistent
Summary:
This is going to be necessary in order for us to directly deal
with the socket path (which cannot change on subsequent reads), which
we need if we want to get rid of PyreQueryHandler.

Why do we want to get rid of PyreQueryHandler? Becasue it will be
dramatically easier to implement a shadow-mode flag in which navigation
queries give instant responses if we can get rid of the queue-based
control flow indirection.

Reviewed By: joshuaxu1995

Differential Revision: D39079398

fbshipit-source-id: 89466e677cf41dc767b4db13daff7eac6727430b