Add a `get_socket_path` method to the server options
Summary:
The immediate benefit is that it reduces repetition a bit, but more
importantly: once we add a shadow mode flag to the StartArguments,
the socket path will no longer be determined by the project_identifier
alone. Wrapping it in a method makes it easy for us to ensure we correctly
deal with the StartArguments flags.

Reviewed By: grievejia

Differential Revision: D39079393

fbshipit-source-id: 59cdc228db27707c173b8ebca99f9b10c3d683b6