Rename `PyreDaemonStartOptions` to `PyreServerOptions`
Summary:
I named this orginally based on the assumption that the options were tied
to the daemon start only, but after some more working with the code I've
realized that several of these options affect the python-side language
server code.

Simply calling this `PyreServerOptions` will do.

The source of my confusion was that these options were originally named
`PyreServerStartOptions`, hence my assumption that they had to do
exclusively with starting the daemon; the new name is clearer (since one
of the really key things is that we attempt to reload these options whenever
the configuration changes!)

Reviewed By: joshuaxu1995

Differential Revision: D39079395

fbshipit-source-id: 718e8a35de7757cfa55366fddb97c73be3dbd0ce