Get rid of quite a lot of test boilerplate
Summary:
Somehow it seems like we got into some extreme copypasta-happy
coding patterns, copying the same set of irrelevant arguments over
and over again in our tests.

This isn't necessary, let's not do it. Just set some default arguments
so that we only set the options that are important for a given test.

Reviewed By: joshuaxu1995

Differential Revision: D39079399

fbshipit-source-id: 3383de62f6ab797194284311f658ec5eaa63885e