Move `write_telemetry` into `PyreQueryHandler`
Summary:
This reduces the number of arguments we are passing around.

This commit may seem odd in a series where I'm working toward
getting rid of PyreQueryHandler... why would I be moving more logic
*into* it?

The answer is that while I do want to kill PyreQueryHandler - or at least the
fact that it's in a background task - we're still going to have some object
(possibly PyreServer itself but I hope not, since we kind of want to break that
up too) that knows how to run queries and write telemetry change is still
useful.

More importantly, I need to clean up the interface a bit before I can
actually perform the refactor, so even if we wind up splitting this back
out again it is going to help me in the short run

Reviewed By: grievejia

Differential Revision: D39079396

fbshipit-source-id: 6e3ab7870acaf661b9c85aa688473b96fca243b3