For now, do not pass unnecessary flags to PyreQueryHandler
Summary:
The next stage of my refactoring of `persistent.py` code needs to be
getting rid of the `pyre_query_manager` task manager and the queue
of requests; instead want to turn the `PyreQueryHandler` into an
ordinary class that knows how to deal with requests and responses,
but without it being in a background task or using a queue.

That state will be temporary, but to get there I need to simplify
the control flow by getting the flags that control behavior from
`self` instead of passing them in.

-------------

This is in some ways the *opposite* of the desired end state, which
is to flatten all the logic and kill `PyreQueryHandler` entirely
(and grievejia has expressed dissatisfaction with object-oriented
apis like this that set multiple flags and use only some of them
in methods, because that can make testing harder).

Those are fair critiques, but right now I need to put the code in
a state where the PyreServer can call into a reworked PyreQueryHandler
directly, and that means I cannot be passing in flags which - as of
the *current* code state - the PyreServer does not have access to.

We can debate how to make all flag passing simple later once things
are in a more sensible state, for now we need to be optimizing
for the ability to gradually refactor.

Reviewed By: grievejia

Differential Revision: D39079394

fbshipit-source-id: 5e3a4a608b8818d97734389302ff6aae11f66df6