For type coverage, pull responding / logging up to server
Summary:
At this stage we are almost ready to implement shadow mode in
`persistent.py`. But I still think the breakdown of responsibility
between PyreServer and the RequestHandler is messy, which makes it
hard to see where to fit shadow mode.

I want to do a couple more refactors in an effort to strip RequestHandler
down to the bare minimum. The biggest change I want to make is to
stop sending data directly to clients from RequestHandler, I want its
methods to act as pure functions, and let PyreServer deal with all client
interactions.

This commit kicks off that refactor, by changing the type coverage query to
to be handled in this way. I'll be flipping the queries over one at a time
to keep the diff sizes manageable, but all of them will look sort of like this:

- the RequestHandler should take specific arguments rather than a dataclass (I
  don't see much benefit to the extra layer of types, I want to get rid of
  this)
- RequestHandler will return an optional lsp response - handling the conversion
  from the server response type
- RequestHandler will *not* send data over stdout to the editor, that will be
  handled by PyreServer

In the tests, the related changes are that:

- json output tests migrate to the PyreServer method tests rather than the
  RequestHandler tests (in this case I had to add it, we were actually missing
  the json output entirely)
- I inject optional mock data into the MockRequestHandler; this is hacky but
  works okay I think; the resulting tests are pretty clear which is my main
  concern right now

Reviewed By: grievejia

Differential Revision: D39108952

fbshipit-source-id: 155d52f7cb94a3865b0970e5d8bba1d08be74d01