Move server options into the ServerState
Summary:
In theory, a config reload could change the settings that control
server behavior.

I don't consider this to be important, we could just ask the client to
reload the window. But given that the deamon manager background task
reloads config before restarting the daemon anyway, and given that we
already allow this background task to modify other state fields, I think
it makes sense to just put the server options in the state.

Reviewed By: joshuaxu1995

Differential Revision: D39135691

fbshipit-source-id: 235c9687a4772d1c21c78903033556fb643758d1