Fix lsp names: Diagnostic should have an LspRange
Summary:
The Diagnostic type in type coverage results is an editor construct.
The ranges and positions should be (and in fact are) using editor
position information, not daemon-style position information.

The types here were incorrect; it didn't matter because we weren't
actually using the conversion tools, and the shapes of the types are
the same, but it was nontheless a bug and extremely confusing to read.

Reviewed By: joshuaxu1995

Differential Revision: D39136245

fbshipit-source-id: 55bc0cec4253abec08fd673e1e39e5f1695aed35