Fix lsp names: PyrePosition, PyreRange
Summary:
It seems like the language_server_protocol class names have
gotten into a surprisingly bad state. I've already renamed several
classes that had cargo cult nonsense in the middle of the name to
be more similar to the LSP protocol, but we still have multiple
issues.

The most severe problem is that the names aggressively lie to us.
The semantic type `lsp.Position` is *not* an lsp position, it is
supposed to represent a Pyre position, where the daemon uses a
different line indexing scheme. Similarly for `lsp.Range`.

But the names are sufficiently confusing that not only do I get
frustrated, but we've written *multiple* bugs where we incorrectly
use `Position` and `Range` in language server protocol classes.
As a result, the names have become meaningless and there is actually
zero type safety for normal code!

In this commit, I rename these classes to `PyreRange` and `PyrePosition`,
which exposes the fact that in several locations we're actually using
them improperly; I'll clean those up separately.

Reviewed By: joshuaxu1995

Differential Revision: D39135690

fbshipit-source-id: e7d1fc5f528c6d964cacd6753c7836480c0e3b75