Add a debug step to the pyre workflow
The version of pyre-check-nightly that is failing to run in CI
(not just for us, but also for torchx) appears to work just fine
on my machine. I think I need to get in and poke around to
determine what is happening.

There may be other ways to debug, but if this works and I can get
an interactive shell inside the instance then that will probably
be a faster way to debug.