Add tests verifying ModuleTracker invalidation
Summary:
Jia has reported a bug in ModuleTracker's lazy finder invalidation.

The problem appears when we add a new module to a project; at least
in some cases it appears that we fail to invalidate stale directory
reads, which causes the new modules to be unloadable.

This commit verifies that
- in general this does work, by taking the existing invalidation
  test based on removal and tweaking it for additions
- there seems to be a special-case bug in the behavior for children
  of the root: the exact same test but with the modules directly at
  the top-level fails.

Reviewed By: pradeep90

Differential Revision: D39160709

fbshipit-source-id: d267efd7448e24e6250bd00480953932c5f3a396