Extract this-and-all-parents logic from ModuleTracker
Summary:
The root of some bad cache invalidation logic is that the code to
get "this reference and all parents" was embedded directly into the
code, which made it easy to not notice an edge case: the prefixes
function explicitly excludes the empty reference, which was desirable
for the case of UnannotatedGlobalEnvironment's "get all possible qualifiers"
logic, but is a bug for lazy module finder invalidation.

Extracting the code as-is into a better-named function (and making the
`prefixes` function private) and adding unit tests makes it more
obvious that the bug is here: the empty reference is missing from the
results except when the input is itself exactly the empty reference.

I'll fix the bug in the next commit

Reviewed By: pradeep90

Differential Revision: D39160708

fbshipit-source-id: 6206071a32087d36f579faffc2342c538f2f62c8