Add resulting module type for lattice of maps.
Summary:
Problem: So far, I'd only tested the generated functions in the resulting module of `LatticeOfMaps.Make`. However, the existing functions don't work because we're not exposing the original types `t` and `key` :P

Want to specify the constraint that the resulting module has the same `t` and `key` as the input module.

Reviewed By: stroxler

Differential Revision: D39183162

fbshipit-source-id: 5267fa9419b6ecdd41dfaedcc6c7fa5c0e0e815b