Factor out functor for generating `less_or_equal`, etc.
Summary:
Problem: For readonly analysis, I want the lattice type to be a map from variables to their readonlyness. I'd need to write `join`, etc. for this data structure.

`typeCheck` uses `Refinement.t`, which uses a private `MapLattice` module that generates `join`, etc. for the map using the supplied `join_one` for the values. I want to reuse that.

This stack copies that over into a separate file so that it can be reused. I'll add in the lattice functions one by one in coming diffs. Once I'm done, I'll use this in the original and DRY things out.

This diff: Copy and test `less_or_equal`.

Next diff: `join`.

Reviewed By: stroxler

Differential Revision: D39183160

fbshipit-source-id: b9e4e29c2b82e05fa4e29fcea08799dad3cd044e