Extract response wrapping boilerplate in persistent tests
Summary:
The wrapping of a raw json-able object into a SuccessResponse
and dumping to a string is repeated in many of the PyreServer
tests, and it generates a bit of line noise (which gets worse
with a refactor I'm looking to make next).

Let's extract it into a helper, which removes two layers of
nesting from all the callsites, making it more obvious what
we really care about.

Reviewed By: pradeep90

Differential Revision: D39210516

fbshipit-source-id: 0870164b3da837250092dcfb748e2f069e234555