Fix a telemetry logging json serialization bug
Summary:
Json serialization isn't even close to type-safe in Python, we have to
cast to object in most places. As a result, it's very easy to accidentally
break the telemetry logging because there's no test coverage.

I was pretty sure we would break this without some tests, turns out I already did :(

I knew the lack of test coverage was dangerous already; I actually caught
this working on a stack (see the subsequent commits) to make sure telemetry
logging was run in the unit tests.

Reviewed By: pradeep90

Differential Revision: D39222470

fbshipit-source-id: ee537c34fc2d203fc35c777d9ad442317917d7dd