Implement caching (#653)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/653

Implements caching in pysa playground to return the same response when a same model and code is being used.

A md5 hash is computed on the concatenated model and code input which serves as a file name for a cache file which stores the output of Pysa for that speicifc pair of input.

Adds code to create cache when we don't have cache and to use cache when we have it.

Signed-off-by: Abishek V Ashok <abishekvashok@fb.com>

Reviewed By: 0xedward

Differential Revision: D38859999

fbshipit-source-id: f47e493fe43611f33952a27d2949ef944ab171a1