Fix the build_pypi script
Summary:
Originally, I had enoded the check here as *disallowing* the
string 'platform' to show up, which is correct but seemed hacky
and this was pointed out in diff review.

So I locked it down more, but by then I had forgotten that macos
builds don't have any dynamic linking, so now the script fails
on macos builds!

I was relying on CI to catch this, but CI only builds pypi on
linux, not on macos. I caught it via the nightly failure emails.

Reviewed By: shannonzhu

Differential Revision: D39249644

fbshipit-source-id: fda7730b40ce2f7597787ce36b60616ad9afef91