Buffer writes to the taint output file
Summary:
The current implementation is very inefficient for big analyses, because it stores all models as json in a big buffer, then write it all in a file with a single call. That buffer can be very large on some codebases, leading to out of memory errors.
In this diff, we simply call `Json.to_channel` followed by a `fprintf "\n"` which should force a flush after every model. This might be slower but will be beneficial for memory usage.
In the long term, we should implement sharding for taint output files.

Differential Revision: D39225761

fbshipit-source-id: 6170d39b85f154fa9b4c539ba1e876a079e6d393