Coverage: Avoid early return once inside branching blocks
Summary:
I know that sometimes we prefer early returns to
following structured programming rules, and I can see that in
python code that requires refinement it is often advantageous.
I'm cool with early returns at the top-level of a function's
branching logic.

But I believe structured programming - in which all branches have
to have equivalent "type information", sort of the equivalent of
what would be forced on us in an expression-blocked language like
ocaml - is *much* easier for the eye to follow; it forces the code's
visual structure to accurately reflect the control flow semantics.

As a result, I'm strongly opposed to the use of early returns except
at funciton top levels. Once we're inside of nested branching
control structures, I very much believe in following the structured
programming principle that either there should be no returns at all,
or all returns should be in leaves of the branching tree.

Reviewed By: grievejia

Differential Revision: D39242508

fbshipit-source-id: eebfdfbd5a67db64e57c22395f5d9413a0a08b42