Coverage: Remove unreachable code
Summary:
The type coverage logic is hard to understand, it is a maze of
what I think are not fully thought through branching conditions
(for example, if the file is not type checked why do we try
to get expression level covarage?)

This commit chips away at the problem: the result of calling
`_make_expression_level_coverage_response` is not optional, so
that None check is dead code. Let's get rid of it.

Reviewed By: grievejia

Differential Revision: D39242509

fbshipit-source-id: 256c70e583488497eb28fb28fdbf21e2cd17599a