Coverage: No expression level coverage for uncovered files
Summary:
I don't think this commit actually changes any behaviors that can
happen in the real world - I doubt that the daemon would consider
a module untracked and yet be able to produce expression-level coverage.

But the control flow pretended this could happen, and we even had
a unit test verifying it.

Let's just be consistent: if a module is untracked, we do not produce
type information.

Reviewed By: grievejia

Differential Revision: D39242507

fbshipit-source-id: ee4192f755bd653ac1fed1e47ad4848bd4ae8971