Convert language server feature availability to an enum
Summary:
I want to get away from using boolean flags to control feature
availability toward using an enum.

The default enum can have two values, enabled and disabled,
but features with a shadow mode will need a third value and
I also want it to be easy to, for example, specify that we're
using an experimental version of an existing feature (e.g.
hover at some point might have a flag to enable "pretty hover"
instead of the plain text we currently render and/or to
add docstring information).

Reviewed By: grievejia

Differential Revision: D39249758

fbshipit-source-id: f174970688ce35e7f3771542dd0fc3b540dba171