Include telemetry writing in unit tests (temporarily, we hope)
Summary:
In this commit, I finally get to including telemetry writing
in test coverage. Turns out I was worried about test coverage
because I knew it would be very easy to accidentally break
telemetry writing in a refactor due to lack of type safety; turns
out I actually had broken it already :(

NOTE: It's maybe not so nice to rely on unit tests of telemetry code,
especially if the main likely source of bugs is json serialization. There may
be a refactor that would make json serialization bulletproof using dataclasses_json,
which would reduce the need for ugly tests. But we want to move forward with
shadow launch as fast as possible so I don't want to get bogged down trying to do
that just yet.

Reviewed By: grievejia

Differential Revision: D39222265

fbshipit-source-id: 45e938c5a7009cfcf43a8b14349b461b6454892d