Functionalize expectations over responses in PyreServer tests
Summary:
I don't really want to hardcode everything about telemetry responses in
unit tests; that's more brittle than we really need, I just want to be
able to sanity check that it serializes and has a few basic properties.

How can we square that with the fact that we do want to fully validate
the content of actual responses that the editor will interpret? We
inject the expectiations via callbacks, and make a helper function so
that it's easy to make an "exactly equals" expectation :)

Reviewed By: grievejia

Differential Revision: D39222272

fbshipit-source-id: a447e318f0d09248629bdf6e5a579be9cccbcb73