Factor out helpers from hover test
Summary:
I'm working on refactoring `persistent_test` to make it
more flexible; I want to get rid of the existing helper
functions like `_assert_hover_response`, which were a big
improvement over the previous tests but are overly coupled
to a model of one-request-one-response.

I want more flexibility for two reasons:
- we need regression tests on telemetry logging; on at least
  one occasion I broke persistent entirely and only caught it
  with a manual end-to-end test, because json serialization is
  not type-safe and there's no unit test coverage
- I want to be able to test shadow-mode behavior, which I'm almost
  ready to roll out, but that would be super weird to test without
  telemetry also being covered

But in order to stop using helpers like `_assert_hover_response`,
I need to have more generic helpers because inlining all the low
level logic into tests lead to code that I couldn't skim, which
is not a state I want to be in.

So, this commit factors out generic helpers for:
- setting up the PyreServer to test against (which is almost
  the same in every case)
- reading the messages sent over the output channel; this logic
  is pretty low-level and tricky because the dict serialization
  order from cached_schema().dump() calls is nondeterministic, but
  it's easy to write in a highly reusable way.

For now I'm only modifying the _assert_hover_response helper to use
the new helpers; in later commits I'll be getting rid of it entirely
to break the hover tests out and applying the same refactors to
other request types.

Reviewed By: joshuaxu1995

Differential Revision: D39222267

fbshipit-source-id: 3b40298cb766b6a8ef7778429e9af26628ed44ef