Remove pre-existing taint output shard files
Summary:
It is common to run pysa with the same results directory multiple times. To prevent confusion, let's remove existing taint output shards before writing new ones. This will prevent issues when the total number of shards is different and some files persist between runs.
Also note that it is not a problem for other files or non-sharded mode, since we will just overwrite those files. It is only a problem if one run wrote `taint-output@xxx-32.json` and the second run writes `taint-output@xxx-33.json`.

Differential Revision: D39325889

fbshipit-source-id: 17a22de1ccadbdf38b0ec8e6f5f783d3e259e742