In tests, use lsp_line + 1
Summary:
The logic is that the daemon line is always 1 more than the
lsp-side line becuase of 0- vs 1-indexing. The code should
show this more clearly.

NOTE: This is a follow-up to a diff review comment; I didn't fix it
in-place because it was likely that restacking would lead to the
issue being re-introduced.

Reviewed By: joshuaxu1995

Differential Revision: D39321764

fbshipit-source-id: bfa06ee2da6d55090f80063ec7ae88c11289d362