Create shadow-mode handling for definition requests
Summary:
This commit adds a second availability enum type that lets us mark
specific features as having a "shadow" mode in addition to "enabled"
and "disabled". It sets that enum as the type for the `definition` feature,
and adds logic to handle when the definition feature is in shadow mode:

- immediately send an empty response
- when we get the actual result, do not send a response
- but do always log telemetry

----

Note that we unfortunately cannot easily reuse the code from `Availability`
in `AvailabilityWithShaodw`. I kind of wish we could use inheritance here
to get code reuse without having type system implications, but that is not
possible and so enums cannot be inherited.

But I think that is a small price to pay for how easily the *interface* can
be preserved when adding shadow mode here. From the CLI's point of view,
nothing changes unless we want to actually use shadow mode.

Note: this actually only works as a non-breaking change because the CLI is
stringly typed. If we were exposing pyre as a library, the fact that I'm rigidly
specifying the enum type means we'd break all callers.

If we ever decide to expose pyre as a library, it would probably be good to
define a type alias for each feature (e.g. DefinitionAvailability) so that
a change like this could be made compatible from the caller perspective.

Reviewed By: grievejia

Differential Revision: D39250127

fbshipit-source-id: 2bc622c8ec42e591640058a5f8e2a855a5e242b8