LanguageServer features in cli: add the cli interface
Summary:
Note: the code is kind of ugly at this point - we pass all these flags around several times,
leading to quite a lot of boilerplate. The driving reason for this is that we're delaying
evaluation of LanguageServerFeatures until the config is read (potentially re-read) off disk.

That won't be necessary once we get rid of these fields from the configuration, which we will
as soon as the vscode plugin is updated to use `persistent` flags and released; at that point
we can not only clean up most of these arguments lists, but we can also get rid of many of
the click options defined on the base `pyre` command. So once we've finished the 3-phase rollout
this will be much closer to a net-neutral lines-of-code change rather than the big addition that
it appears to be in this intermediate state.

Since the arguments list has to repeated a few times (some of the repetition can be removed soon,
but the click-level stuff cannot be), I also decided to alphabetize it.

Reviewed By: grievejia

Differential Revision: D39281300

fbshipit-source-id: a4a5732b1c92d35cdd56d66a228fc40a1392e723