LanguageServer features in cli: decouple the availability types
Summary:
As noted in my previous diff, the big benefit of using an enum
rather than boolean flags to control feature availability in an IDE
is that it allows us to experiment with functionality without breaking
the interface on the cli.

This is the only way we can move fast in a world where the vscode plugin
is deployed independently of pyre itself - breaking changes are vastly
slower to rollout, and nearly guaranteed to break at least some users.

The same problem is much less serious in python because we don't currently
consider the python API to be stable, but nontheless: we'd also like the
ability to rapidly iterate on a specific feature making only local changes
where they are needed.

The direct use of `Availability` and `AvailabilityWithShadow` is
failing us here: if we add an extra experimental setting to any given feature,
we globally have to change the type of the enum. This is exactly what I'm
trying to avoid in the cli, and it makes sense to avoid it in the python type
signatures as well.

So, instead lets define the two concrete enums as private classes, and then
use type aliases so that each feature is decoupled from the others. This allows
us to add new enums on-the-fly as needed to experiment (e.g. with a prettified
hover, which is almost certainly something we'll need if we want to be competitive
with jedi down the road).

Note that type coverage was *already* an availability flag with a divergent enum
type, and we can now clearly see how this pattern plays out: by renaming the coverage
"level" to coverage "availability" it now becomes much more parsimonious with the other
ide features.

Reviewed By: grievejia

Differential Revision: D39281298

fbshipit-source-id: e65e7a55bbd21fc30488a9d3d32da4d3638f63ef