Separate logs and sockets: add PyreFlavor + socket path logic
Summary:
In an earlier, abandoned diff I considered allowing us to name pyre servers.
This was rejected because there is the possibility of a user attempting
to specify a long name that would pass the 104-character limit on socket path
lengths.

The proposed solution was to implement a binary flag, not as a flag but by
hardcoding its binary nature into the very module signature. I decided to reject
this idea for the same reason that I'm rejecting binary flags to control language
server features: it isn't forward-compatible in any reasonable way
with a world with more than 2 options (also, a binary flag is going to have to live
*somewhere* unless we want to make separate entrypoints for classic and shadow
which I decided against doing).

Instead, I'm choosing a similar approach to the one I used for features: we can
restrict to <100 character names by restricting the name to one of a small set
of known identifiers, and add new identifiers as needed later (in particular I
anticipate us having a separate identifier for a code navigation server).

Reviewed By: grievejia

Differential Revision: D39296125

fbshipit-source-id: 4057f321d8e00e59fc8f79806464886e69ac710f