Add New Query Type: ReferencesUsedByFile Query
Summary:
**Context**: This diff is part of a larger diff stack that implements the new query `ReferencesUsedByFile` - which Glean will use to implement find references.

We need to create a new request type called `ReferencesUsedByFile` which takes in a path and figures out the types of all the symbols at that path.

**Diff Details** Nothing super interesting to point out except that `ReferencesUsedByFile` will be implemented similarly to `Types`.

Also, this diff adds all the tests that were in TypesInFiles so that ReferencesUsedByFile doesn't basically suffer a regression: all the tests that operated correctly  or TypesInFiles should return the correct result for ReferencesUsedByFile.

NOTE: this diff doesn't add any of the special tests for the logic we want to test with importing classes from other modules: just previous tests that exist in typesInFiles.

Reviewed By: pradeep90

Differential Revision: D39099708

fbshipit-source-id: 357a9470a445c5ccda918d487a8c81868a8f6b94