Use same command line arguments for no_daemon_query as is used in check
Summary:
**Context**: This diff is part of a diff stack that aims to create a no daemon version of pyre query. See base diff: D39273627

**This diff**: Changes the command line arguments used in pyre no daemon query so that it matches closely with those of the check configuration: https://www.internalfb.com/code/fbsource/[f787cb7fc178caa257adccdf5c59c07f1dfaf2e1]/fbcode/tools/pyre/client/commands/check.py?lines=84

I tried to match the configuration as closely as possible, but there are a few exceptions:

1. memory profiling/profiling output are not the same
2. I didn't include a debug flag because I need to make sure it's implemented in no daemon query first.
3. checked_directory_allowlist/ checked_directory_blocklist are also ignored. Edit: this is addressed in a future diff: https://www.internalfb.com/diff/D39310372

In future steps, I will do more verification on the other fields. The way I plan to do this is to inspect each command argument. If anyone has any input on how to determine which configuration flags are necessary (if you have a quicker approach) that would be greatly appreciated!

Reviewed By: pradeep90

Differential Revision: D39306544

fbshipit-source-id: f3338cab394c454a3631968f604698cf3c45c763