Abstract Rooted Tree
Summary:
Adds a variation of the AbstractTreeDomain that contains a separate domain paired at the root.

   RootDomain x Tree(Element)

The root domain is kept up-to-date as follows. Every selection, etc on the Tree simply leaves the Root domain intact. But if we add a subtree to the tree domain, we join the root domains of the top and the subtree.

Reviewed By: williewillus

Differential Revision: D39336102

fbshipit-source-id: ac842cdb17defb2e882afe068f8e9dae59b838a8