Separate logs and sockets: Use the flavor in get_socket_path
Summary:
As of the previous commit, we're able to spin up a language server
and daemon with a shadow flavor. But I initially missed that the control
flow for the socket path when launching the server versus connecting
to it are separate, and we were still using the default path when
connecting (which led to the language server crashing).

This commit modifies the connection socket path to account for the
flavor as well.

Reviewed By: pradeep90

Differential Revision: D39374996

fbshipit-source-id: 78f87e316b8ba793fafb89baf5c6888d594db739