Encapsulate daemon path logic better
Summary:
Instead of having constants floating in higher-level functions, lets
have low-level functions that specifically handle the dynamic parts
of the logic to get log paths.

This will make it easy to add flavor handling (which we need - different
daemon flavors need to be able to coexist, and therefore need separate
logs).

Reviewed By: grievejia

Differential Revision: D39369849

fbshipit-source-id: d13087e0ac1fce980a81445763ddf57d7e283b84