Separate logs and sockets: use flavor to set client log path
Summary:
Similarly to the previous diff, in order to reasonably begin working
on vscode integration of shadow mode I need the ability to clearly
separate out client logs based on the pyre flavor so that logs from
shadow-mode and classic servers (and later code nav servers) don't
all wind up in the same place.

As with the previous diff, handling the separate log paths in
`pyre rage` and `pyre info` can wait; it is maybe a blocker on
an actual shadow mode launch but I can at least start developing
the logic with just myself in a GK without that code up and running.

Reviewed By: grievejia

Differential Revision: D39370652

fbshipit-source-id: 38191bb203ce6c63fafdbfa242797fcfb382f09a