Make flavor required in `get_socket_path`
Summary:
This rules out bugs where we weren't accounting for the flavor, and as
a result accidentally picked up the `CLASSIC` default value.

There were no bugs per se, but one case that could have become a bug
was `start.py` where we were assuming CLASSIC flavor - which is true
because the entrypoint forces it - but in theory StartArguments
has its own flavor that could be different if we ever add a flag to
set the flavor.

Reviewed By: grievejia

Differential Revision: D39395224

fbshipit-source-id: d81c0bc0710cf21892e8422826a4d38e5232e9a8