Separate logs and sockets: use flavor to set daemon log location
Summary:
In addition to not clobbering the socket paths for different pyre flavors,
we also need to keep the logs separate. This commit focuses on the daemon
logs - it uses the same path suffix logic we use for sockets to set the
daemon log paths.

At some point we'll need to extend `pyre rage` and `pyre info` to handle
this cleanly, but for now I'm focused mostly on what's needed for
persistent so that I can hook up vscode and verify end-to-end behaviors.

Reviewed By: grievejia

Differential Revision: D39369850

fbshipit-source-id: 9aa61b047e2082f3a25f8b80fdb7e6aff4877568