Add New Flag to ScratchProject to Create Environment w/ No Validation Class Lookup Failure
Summary:
**Context**:  This diff is part of a stack that implements the new query `referencesUsedByFile`. See the base diff here: https://www.internalfb.com/diff/D39065278 (https://github.com/facebook/pyre-check/commit/5bc480e3c4e04883f608e05a50c5f61c4d8667dc)

This diff addresses a comment made by Pradeep in the previous diff where he mentioned we could make the `no_validation_on_class_lookup_failures` to be a CLI configuration. For the short term, we need to add a flag to `queryTest.ml` that allows us to toggle the value of that `no_validation_on_class_lookup_failures` flag. This diff adds that flag to the test files.

Reviewed By: grievejia

Differential Revision: D39396188

fbshipit-source-id: f902554b735f3be102e0a72e025e8f4fcd2584a3