Move Tests From Top of File to Appropriate Spot
Summary:
**Context**:  This diff is part of a stack that implements the new query `referencesUsedByFile`. See the base diff here: https://www.internalfb.com/diff/D39065278 (https://github.com/facebook/pyre-check/commit/5bc480e3c4e04883f608e05a50c5f61c4d8667dc)

~~Currently, construction of environment tables occurs upon server start. This is a problem because we want the environment tables to have an additional flag that is a query level flag: the value of `no_validation_on_class_lookup_failure` will depend on the query, and since the query gets received by the server after server start, the very initial environment table may have an out of date value for the flag `no_validation_on_class_lookup_failure`. ~~

~~That suggests that an environment needs to be constructed *after* a query is received by the server, which I'm doing here.~~

~~**This diff**:  As mentioned above, this diff adds logic to the `ReferencesUsedByFile` query that constructs an identical environment to the original environment (in terms of the environmental controls), *except* for the value of the `is_references_used_by_file_query` will be set to true.~~

**This diff**:

This diff does nothing except move tests from the top of the file to the appropriate spot. This diff was originally intended to construct an environment in the query, but we decided against it, for the reasons below.

We previously had the idea to construct a separate environment in the `ReferencesUsedByFile` but we were concerned that it would be a bad solution, leading to new environment table being constructed after every query, which would leak tons of memory. The problem still exists that there is no way for the `no_validation_on_class_lookup_failure` flag to be true for only `ReferencesUsedByFile` queries, but as mentioned in the diff comments, we will incorporate this flag into the `noDaemonQueryCommand`.

The only remaining change after we deleted the changes above was to move tests from the top of the file to the appropriate spot, and those changes will remain in this diff.

Reviewed By: pradeep90

Differential Revision: D39273631

fbshipit-source-id: 72045fef1ba33a5edc858ef6ccfd3dfb5d50b300