Demonstrate the wrong order of applying tito sanitizer on arguments when combined with taint transforms
Summary:
When the body of a function is inferred as having a taint transform, and the
function is annotated with a tito sanitizer on its argument, the sanitizer
should be applied to the result after applying taint transform, as opposed to the
argument (that is annotated with the sanitizer). The later this is the current
implementation, which means first applying the sanitizer to the argument and
then applying the transform.

This essentially reproduces the confusion in D38533103.

The new integration tests here enumerate the syntactic locations that
annotation `Sanitize[TaintInTaintOut]` can be used.

Reviewed By: arthaud

Differential Revision: D39286424

fbshipit-source-id: f450d13e01fe5dd1a340911696772b9963706701