Disable taint tree invariant checks in production
Summary:
This diff disables the taint tree invariant checks when the `PYSA_CHECK_INVARIANTS` environment variable is not set.
This should improve performance in production by disabling safety checks.
This diff also adds a rule in dune that defines `PYSA_CHECK_INVARIANTS` in tests so that the checks are still used.

Reviewed By: dkgi

Differential Revision: D39528356

fbshipit-source-id: 876d653dfc44d4f4599b4d0f418eb754e79c2669